Added a get route and registration route to controllers
