+++ server/controllers/registrationRoutes.ts
@@ -7,11 +7,6 @@ const registrationRoutes = express.Router();
 
 registrationRoutes.post('/sign-up', registrationController.createUser);
 
-registrationRoutes.delete('/delete-account', async (req: Req, res: Res, next: (para1: any) => any) => {
-  const user = await User.destroy({
-    truncate: true
-  });
-  res.status(204).json(user);
-})
+registrationRoutes.delete('/delete-account', registrationController.deleteUser)
 
 module.exports = registrationRoutes;
